+++ minicluster/src/main/java/org/apache/accumulo/cluster/standalone/StandaloneClusterControl.java
@@ -151,7 +151,7 @@ public class StandaloneClusterControl implements ClusterControl {
     String[] cmd = new String[] {serverCmdPrefix, accumuloPath, Admin.class.getName(), "stopAll"};
     // Directly invoke the RemoteShell
     Entry<Integer,String> pair = exec(master, cmd);
-    if (0 != pair.getKey().intValue()) {
+    if (0 != pair.getKey()) {
       throw new IOException("stopAll did not finish successfully, retcode=" + pair.getKey() + ", stdout=" + pair.getValue());
     }
   }
@@ -170,7 +170,7 @@ public class StandaloneClusterControl implements ClusterControl {
     String master = getHosts(MASTER_HOSTS_FILE).get(0);
     String[] cmd = new String[] {serverCmdPrefix, accumuloPath, SetGoalState.class.getName(), goalState};
     Entry<Integer,String> pair = exec(master, cmd);
-    if (0 != pair.getKey().intValue()) {
+    if (0 != pair.getKey()) {
       throw new IOException("SetGoalState did not finish successfully, retcode=" + pair.getKey() + ", stdout=" + pair.getValue());
     }
   }
