+++ test/src/main/java/org/apache/accumulo/test/VolumeIT.java
@@ -161,7 +161,7 @@ public class VolumeIT extends ConfigurableMacBase {
       assertEquals(4, fileCount);
       List<DiskUsage> diskUsage = connector.tableOperations().getDiskUsage(Collections.singleton(tableName));
       assertEquals(1, diskUsage.size());
-      long usage = diskUsage.get(0).getUsage().longValue();
+      long usage = diskUsage.get(0).getUsage();
       log.debug("usage {}", usage);
       assertTrue(usage > 700 && usage < 800);
     }
